added cache strategy
now we have a map (or dictionary) containing all the cached images and a counter for the times used.

when we have to cache a 6th image (we setted MAX_CACHE_SIZE to 5, in any case it would be MAX_CACHE_SIZE + 1) we delete and remove from our dictionary the least used image.