receive some parameters from argsv
with this changes we can implement our new cache management structure.

You have to clarify in the parameters when running the server the amount of images you want to cach at maximum and the server hostname and port where you want to get the images from if you don't have the images cache.